refactored track specific elements to their own package
